--- app/src/main/java/com/techcourse/controller/LoginViewController.java
@@ -1,27 +1,0 @@
-package com.techcourse.controller;
-
-import context.org.springframework.stereotype.Controller;
-import jakarta.servlet.http.HttpSession;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import web.org.springframework.web.bind.annotation.RequestMapping;
-import web.org.springframework.web.bind.annotation.RequestMethod;
-import webmvc.org.springframework.web.servlet.ModelAndView;
-import webmvc.org.springframework.web.servlet.view.JspView;
-
-@Controller
-public class LoginViewController {
-
-    private static final Logger log = LoggerFactory.getLogger(LoginViewController.class);
-
-    @RequestMapping(value = "/login/view", method = RequestMethod.GET)
-    public ModelAndView loginView(final HttpSession session) {
-        return UserSession.getUserFrom(session)
-                .map(user -> {
-                    log.info("logged in {}", user.getAccount());
-                    return "redirect:/index.jsp";
-                })
-                .map(it -> new ModelAndView(new JspView(it)))
-                .orElse(new ModelAndView(new JspView("/login.jsp")));
-    }
-}
